KhRegularSingers now stores KhRegularSinger objects in a QList<KhRegularSinger>
Previously KhRegularSingers was based on boost shared pointers and
std::vector.  Since we're using QObject based classes for this stuff I'm
in the process of switching over to QLists of objects since it handles
deletion of the objects itself.  Makes the code look a lot cleaner too
by avoiding all of the boost::shared_ptr<blah<blah> > stuff.