Fixing segfault on song insert to regular tracked singer
Fixed a segfault on song insert for singer tracked as regular.  There is
still a major problem which is preventing the adds/moves from being
written out to the database, so they're lost on program restart.  I need
to tear apart the current regular songs/singers object code and re-work
it completely.  I made some very poor design choices the first time
around which got worse as crap was piled on top of them or changed.