Kh - Audio backend now configurable in program
You can now compile in support for multiple backends and choose which to
use in settings.  I plan to make this runtime switchable, but for now it
requires a program restart.

As part of the change, KhAbstractAudioBackend now has its own State type
instead of just relying on QMediaPlayer::State.  The program is no
longer dependent upon Qt Multimedia when not compiling in QMediaPlayer
support.