Bm - Adding missing features, cleaning up a bit
This version breaks compatibility with the previous database schema, if
you were running it previously you will need to delete the
breakmusic.sqlite file and let the program rebuild it from scratch.

I am considering BreakMusic pretty much feature complete at this point.
It isn't intended to be a full featured media player, just a convenient
companion for playing break music with KaraokeHost.  For this program it
will just be small tweaks and bug fixes for a while as I focus back on
Kh.