Ticker tweaks
Make ticker text stick to the top of the widget.  This prevents needing
crazy big ticker areas when trying to bring the ticker text up to the
viewable portion of TV screens.

Reduce overall string length on the ticker by shortening some of the
static elements. "Current:" vs "Current Singer:" and the like.