KH - Work on CDG Window placement/adjustment
Got all of the plumbing and GUI stuff done for fullscreen CDG window
positional adjustments.  hOffset and vOffset don't seem to be working on
Linux, needs investigation.

Fixed a bug that crept in at some point that kept the CDG window from
going fullscreen on the proper monitor on Fedora 23 (Linux).