Kh - More work on KhQtMultimediaBackend
All functionality is now working in the new backend with the exception
of key changing.  Still beating my head against that one.  Pacemaker is
kind of a pain in the ass to use with Qt's audio classes.  I did get a
subclass of QIODevice to basically act as a proxy to the QBuffer that
holds the audio that the QAudioOutput consumes, so there is progress at
least.