Kh - QtMultimedia backend is feature complete!
Still a lot of tweaking and cleanup to do, but at this point it is the
only fully functional audio backend.  The pitch shifter needs some
serious tweaking to get it performant and avoid the occasional buffer
underrun, but I've figured out the plumbing, which was the hard part.

Also, a lot of other random code changes where I'm trying some different
ways of doing things out.  If they don't work out, I'll revert.