Adding new background image slideshow feature
In settings, you can set a directory and it will loop over all files in
that directory, detect whether each one is a valid image file, and loop
over the valid images.  The loop is done alphabetically (case
insensitive), so if you want to control the order, just prefix each name
with 00, 01, 02, 03 or some such.  Currently the time between switches
is hard coded at 15sec.  I'll put in the UI elements and plumbing to
make that configurable later.