Initial gstreamer based recording replacement class
Switching to gstreamer due to QAudioRecorder not supporting any codecs
beyond PCM/RAW on Windows 8.1 or above.  Keeping the old class code
around for now in case Qt starts supporting audio properly on Windows in
the near future as that class is much cleaner and easier to deal with.

Also, cleaned up some random cruft and debug output elsewhere.