Added fair/bottom/next option for manual singer adds
Now there is a combo box on the main form beside the singer name entry.
It allows setting where the singer will be added in the rotation.  Fair
- one full rotation from when they signed up (before the current singer
sings again next rotation).  Bottom - At the end of the current
rotation.  Next - Obviously, after the current singer.

Previous behavior was just to add at the bottom.  This brings manual
singer adds in line with what is available from the requests and
regulars dialogs.