Move all request server functionality to the songbook API class and make songbook API network calls async
Cleaned up all of the json API stuff that was still living in the
request table model and moved it to the songbook API class.  Also
removed the hacky blocking network calls and made it work the way it is
intended in Qt with the exception of db updates, which kinda need to be
blocking.