Fix crash on second rt click after first preview in requests dialog
Program was crashing after you previewed a file and then right clicked
on another.  This is related to a change that was made to preview dialog
to fix a defect. The preview dialog deletes itself on close now rather
than sitting in memory all the time. The main window had been fixed to
work with the new dialog code, but I missed changing the code in the
incoming request dialog's search entry context menu.