Initial stab at auto-advance mode for karaoke
Github feature request issues #23 and #18

When a song ends, it will advance through the rotation from the current
position until it hits a singer with an unplayed song in their queue,
when it does, it will play it after a configurable number of seonds
(default 30).  If it doesn't find any unplayed songs, it just stops.