Prevent break music auto-start from hanging startup on missing files
If break music auto-start was enabled and the mount point (or drive, for
the Windows side) which contained the files in the playlist was
unavailable, the program would loop infinitely trying to play songs and
hang program startup.  This change makes auto-start bail out if the first
song is missing to avoid this possibility.